fix: placeId 가져오는 기능 훅으로 분리
