+++ src/api/shop/entity.ts
@@ -47,25 +47,6 @@ export interface FetchShopResponse {
   totalRating:number;
 }
 
-export interface ShopPinResponse {
-  businessDay:string;
-  category:string;
-  formattedAddress:string;
-  formattedPhoneNumber:string;
-  lateReviewDate:string;
-  lat:number;
-  lng:number;
-  name:string;
-  openNow:boolean;
-  photos:string[];
-  placeId:string;
-  ratingCount:number;
-  scrap:boolean;
-  shopId:number;
-  todayBusinessHour:string;
-  totalRating:number;
-}
-
 export interface FetchShopsResponse {
   shopQueryResponseList: Shop[];
 }
