+++ src/pages/Home/components/Map/index.tsx
@@ -1,8 +1,7 @@
 import useMediaQuery from 'utils/hooks/useMediaQuery';
-import { useEffect, useState } from 'react';
+import { useEffect } from 'react';
 import { useFilterFriend, useFilterNearby, useFilterScrap } from 'store/filter';
 import { useLocation } from 'store/location';
-import { fetchShops } from 'api/shop';
 import SideNavigation from 'components/common/SideNavigation';
 import BottomNavigation from 'components/common/BottomNavigation';
 import styles from './Map.module.scss';
@@ -10,9 +9,9 @@ import MobileOptions from './components/MobileOptions';
 import useNaverMap from './hooks/useNaverMap';
 import useMarker from './hooks/useMarker';
 import useFilterShops from './hooks/useFilterShops';
+import usePlaceId from './hooks/usePlaceId';
 
 export default function Map(): JSX.Element {
-  const [placeId, setPlaceId] = useState<string>('');
   const { isMobile } = useMediaQuery();
   const { location } = useLocation();
   const map = useNaverMap(location?.latitude, location?.longitude);
@@ -26,25 +25,15 @@ export default function Map(): JSX.Element {
   });
 
   const { selected } = useMarker({ map, filterShops });
+  const { placeId } = usePlaceId({
+    title: selected?.getTitle(),
+    lat: selected?.getPosition().y,
+    lng: selected?.getPosition().x,
+  });
 
   useEffect(() => {
     refetch();
   }, [filterFriendState, filterScrapState, filterNearbyState, refetch]);
-
-  useEffect(() => {
-    const getPlceId = async () => {
-      const shops = await fetchShops(
-        {
-          keyword: selected?.getTitle() || '',
-          location: { lat: selected?.getPosition().y, lng: selected?.getPosition().x },
-        },
-      );
-      if (shops.data.shopQueryResponseList.length !== 0) {
-        setPlaceId(shops.data?.shopQueryResponseList[0].placeId || '');
-      }
-    };
-    getPlceId();
-  }, [selected]);
 
   return (
     <div>
